requirements.txt
APScheduler==3.9.1
feedparser==6.0.10
pymongo==4.3.2
opengraph-parse==0.0.6
transformers==4.23.1
python-dotenv==0.20.0
tensorflow-cpu==2.9.0